modified calculators to use doubles, added numotu and fontsize parameters to heatmap as well as adding more sorting options. fixed bug in mothur.h setfilepositions that caused accession names to be truncated when windows line endings were used.

+++ bootstrap.cpp
@@ -16,8 +16,8 @@ EstOutput Bootstrap::getValues(SAbundVector* rank){
 		//vector<double> bootData(3,0);
 		data.resize(1,0);
 		double maxRank = (double)rank->getMaxRank();
-		int sampled = rank->getNumSeqs();
-		int sobs = rank->getNumBins();
+		double sampled = rank->getNumSeqs();
+		double sobs = rank->getNumBins();
 
 		double boot = (double)sobs;
